Add files via upload
This repository contains all the files for the paper “Deep Learning TCP for Improving mmWave 5G Performance due to NLoS Impairments”. It includes python code, ns3-mmWave codes (scenarios and the codes for the protocol), and the inputs for training and testing. One can run the training codes on colab.research.google or python platforms, and the C++ codes can be implemented in the ns3-mmWave module.